adds sluggable model concerns
to use slugs on a model add a slug column to its table and then 'include Sluggable' in the model
