+++ app/models/industry.rb
@@ -1,4 +1,6 @@
 class Industry < ApplicationRecord
+  include Sluggable
+
   before_validation :generate_slug
 
   validates :name, presence: true,
@@ -10,22 +12,4 @@ class Industry < ApplicationRecord
   has_many :categories
   has_many :services, through: :categories
   has_many :pros, foreign_key: :user
-
-  def to_param
-    slug
-  end
-
-  def generate_slug
-    self.slug = name.parameterize
-  end
-
-  def self.find(input)
-    # monkey patches .find to be able to find by id and slug
-    if input.to_i != 0
-      super
-    else
-      find_by_slug(input)
-    end
-  end
-
 end
