+++ app/models/service.rb
@@ -1,4 +1,6 @@
 class Service < ApplicationRecord
+  include Sluggable
+
   before_validation :generate_slug
 
   validates :name, presence: true
@@ -9,21 +11,4 @@ class Service < ApplicationRecord
   belongs_to :category
 
   has_many :projects
-
-  def to_param
-    slug
-  end
-
-  def generate_slug
-    self.slug = name.parameterize
-  end
-
-  def self.find(input)
-    # monkey patches .find to be able to find by id and slug
-    if input.to_i != 0
-      super
-    else
-      find_by_slug(input)
-    end
-  end
 end
